Move card routes to card_routes.py
